requirements.txt
hashids==1.3.1
python-dotenv==1.0.1
python-telegram-bot==21.10
